Separated Entity guarding schemas to folder
