+++ app/domain/entities/document/schemas/tag-collection.schema.ts
